LayoutTests/imported/w3c/web-platform-tests/content-security-policy/frame-src/frame-src-self-unique-origin-expected.txt

CONSOLE MESSAGE: Refused to load data:text/html,blocked_iframe because it does not appear in the frame-src directive of the Content Security Policy.
The origin of an URL is called "unique" when it is considered to be different from every origin, including itself. The origin of a data-url is unique. When the current origin is unique, the CSP source 'self' must not match any URL.



Harness Error (TIMEOUT), message = null

TIMEOUT Iframe's url must not match with 'self'. It must be blocked. Test timed out
